Rearrange files to promote future versioning
