+++ streamlit_camouflage/v1/endpoints.py
@@ -1,47 +1,35 @@
-from fastapi import FastAPI, HTTPException, UploadFile, Response
-from streamlit_camouflage.objects import Clothing, Image
+from fastapi import APIRouter, HTTPException, UploadFile, Response
 from typing import List
 from colorsys import rgb_to_hsv
 import logging
-import uvicorn
-from pydantic import BaseModel
 
-from streamlit_camouflage.fuzzy_classifier import GetValidMatches, GetColorDesc
+from streamlit_camouflage.v1.objects import Clothing, Image
+from streamlit_camouflage.v1.models import Colors, Matches, Outfit
+from streamlit_camouflage.v1.fuzzy_classifier import GetValidMatches, GetColorDesc
 
-app = FastAPI()
+router = APIRouter()
 
 logger = logging.getLogger(__name__)
 
-logger.info("test")
-
-class Color(BaseModel):
-    r: float
-    g: float
-    b: float
-    pct: float
-    name: str
-
-class Colors(BaseModel):
-    colors: List[Color]
-
-class Outfit(BaseModel):
-    outfit: List[Colors]
-
-class Matches(BaseModel):
-    matches: List[str]
-
-@app.get("/")
+@router.get("/")
 async def root():
     return {"message": "Hello World"}
 
 
-@app.post("/rembg", responses={200: {"content": {"image/png": {}}}}, response_class=Response)
+@router.post("/rembg", responses={200: {"content": {"image/png": {}}}}, response_class=Response)
 async def rembg(file: UploadFile):
     try:
+        # Get image contents
         contents = file.file
+
+        # Put the contents in an Image object
         image = Image(contents)
+
+        # Remove the background
         image_rembg = image.rembg()
-        response = Response(content = image_rembg, media_type="image/png")
+
+        # Build the response
+        response = Response(content=image_rembg, media_type="image/png")
         contents.close()
         return response
     except Exception as e:
@@ -50,20 +38,26 @@ async def rembg(file: UploadFile):
         raise HTTPException(status_code=500, detail="Unable to extract colors")
 
 
-@app.post("/colors")
+@router.post("/colors", response_model=Colors)
 async def colors(file: UploadFile):
     try:
+        # Get image contents
         contents = file.file
+
+        # Put the contents to a clothing object
         clothing = Clothing(contents)
+
+        # Get color information
         clothing.extract_colors()
         colors = clothing.colors
         color_names = clothing.get_color_names()
-        response = {
-            "colors": [
-                {'r': rgb[0], 'g': rgb[1], 'b': rgb[2], 'pct': pct, 'name': color_names[i]}
-                for i, (rgb, pct) in enumerate(colors.items())
-            ]
-        }
+        color_dicts = [
+            {'r': rgb[0], 'g': rgb[1], 'b': rgb[2], 'pct': pct, 'name': color_names[i]}
+            for i, (rgb, pct) in enumerate(colors.items())
+        ]
+
+        # Build the response
+        response = Colors(colors=color_dicts)
         contents.close()
         return response
     except Exception as e:
@@ -72,31 +66,25 @@ async def colors(file: UploadFile):
         raise HTTPException(status_code=500, detail="Unable to extract colors")
         
 
-@app.post("/matches")
+@router.post("/matches", response_model=Matches)
 async def matches(outfit: Outfit):
-    logger.info(outfit)
     try:
+        # Get the HSV values for the primary colors
         primary_rgbs = [c.colors[0] for c in outfit.outfit]
-        logger.info(primary_rgbs)
         primary_rgb_norms = [(rgb.r/255.0, rgb.g/255.0, rgb.b/255.0) for rgb in primary_rgbs]
-        logger.info(primary_rgb_norms)
         primary_hsv_norms = [rgb_to_hsv(r, g, b) for r, g, b in primary_rgb_norms]
-        logger.info(primary_hsv_norms)
         primary_hsvs = [(h*360.0, s*100.0, v*100.0) for h, s, v in primary_hsv_norms]
-        logger.info(primary_hsvs)
 
+        # Get a description of the color, in tuples of (TONE, TEMP)
         outfit_color_descs = [GetColorDesc(hsv) for hsv in primary_hsvs]
-        logger.info(outfit_color_descs)
 
+        # Get matches for the given colors
         matches = GetValidMatches(outfit_color_descs)
-        logger.info(matches)
-        return matches
+
+        # Build the response
+        response = Matches(matches=matches)
+        return response
     
     except Exception as e:
         logger.error(e)
         raise HTTPException(status_code=500, detail="Unable to find matches")
-
-    
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.INFO)
-    uvicorn.run(app, host="localhost", port=80)
